+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+Parsifal 2.0.4 (2021-09-10)
+===========================
+
+Bugfixes
+--------
+
+- Improve quality assessment table performance. (#76)
+- Improve data extraction table performance. (#78)
+
+
 Parsifal 2.0.3 (2021-09-08)
 ===========================